# backend/openapi.json
{
  "components": {
    "schemas": {
      "BacktestRequest": {
        "properties": {
          "fees": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Fees",
            "type": "number"
          },
          "initial_cash": {
            "default": 10000.0,
            "exclusiveMinimum": 0.0,
            "title": "Initial Cash",
            "type": "number"
          },
          "limit": {
            "default": 5000,
            "maximum": 5000.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "slippage": {
            "default": 0.0,
            "minimum": 0.0,
            "title": "Slippage",
            "type": "number"
          },
          "strategy": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Strategy",
            "type": "string"
          },
          "symbol": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "default": "1m",
            "maxLength": 10,
            "minLength": 2,
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "strategy"
        ],
        "title": "BacktestRequest",
        "type": "object"
      },
      "BacktestResponse": {
        "properties": {
          "candle_count": {
            "title": "Candle Count",
            "type": "integer"
          },
          "end_datetime": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Datetime"
          },
          "final_value": {
            "title": "Final Value",
            "type": "number"
          },
          "initial_cash": {
            "title": "Initial Cash",
            "type": "number"
          },
          "max_drawdown": {
            "title": "Max Drawdown",
            "type": "number"
          },
          "sharpe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe Ratio"
          },
          "start_datetime": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Datetime"
          },
          "strategy": {
            "title": "Strategy",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          },
          "total_return": {
            "title": "Total Return",
            "type": "number"
          },
          "total_trades": {
            "title": "Total Trades",
            "type": "integer"
          }
        },
        "required": [
          "symbol",
          "timeframe",
          "strategy",
          "start_datetime",
          "end_datetime",
          "candle_count",
          "initial_cash",
          "final_value",
          "total_return",
          "max_drawdown",
          "sharpe_ratio",
          "total_trades"
        ],
        "title": "BacktestResponse",
        "type": "object"
      },
      "Candle": {
        "properties": {
          "OPEN": {
            "title": "Open",
            "type": "number"
          },
          "close": {
            "title": "Close",
            "type": "number"
          },
          "datetime": {
            "format": "date-time",
            "title": "Datetime",
            "type": "string"
          },
          "fecha_carga": {
            "format": "date-time",
            "title": "Fecha Carga",
            "type": "string"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "origen": {
            "title": "Origen",
            "type": "string"
          },
          "spread": {
            "title": "Spread",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "tickvol": {
            "title": "Tickvol",
            "type": "integer"
          },
          "volume": {
            "title": "Volume",
            "type": "integer"
          }
        },
        "required": [
          "datetime",
          "symbol",
          "OPEN",
          "high",
          "low",
          "close",
          "tickvol",
          "volume",
          "spread",
          "origen",
          "fecha_carga"
        ],
        "title": "Candle",
        "type": "object"
      },
      "CandleWindow": {
        "properties": {
          "candles": {
            "items": {
              "$ref": "#/components/schemas/Candle"
            },
            "title": "Candles",
            "type": "array"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "timeframe",
          "candles",
          "next_cursor",
          "has_more"
        ],
        "title": "CandleWindow",
        "type": "object"
      },
      "ClientEvent": {
        "properties": {
          "kind": {
            "enum": [
              "api_failure",
              "unhandled_error",
              "unhandled_rejection"
            ],
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "maxLength": 1000,
            "title": "Message",
            "type": "string"
          },
          "path": {
            "maxLength": 500,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "message",
          "path"
        ],
        "title": "ClientEvent",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InvalidStrategyParameters": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "strategy": {
            "title": "Strategy",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "invalid_strategy_parameters",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "detail",
          "strategy"
        ],
        "title": "InvalidStrategyParameters",
        "type": "object"
      },
      "ServiceUnavailable": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "service_unavailable",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "detail"
        ],
        "title": "ServiceUnavailable",
        "type": "object"
      },
      "StrategyDefinition": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/StrategyParameterDefinition"
            },
            "title": "Parameters",
            "type": "array"
          }
        },
        "required": [
          "name",
          "label",
          "description",
          "parameters"
        ],
        "title": "StrategyDefinition",
        "type": "object"
      },
      "StrategyParameterDefinition": {
        "properties": {
          "default": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "string"
              }
            ],
            "title": "Default"
          },
          "kind": {
            "enum": [
              "integer",
              "number",
              "boolean",
              "string"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "maximum": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maximum"
          },
          "minimum": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "label",
          "kind",
          "default"
        ],
        "title": "StrategyParameterDefinition",
        "type": "object"
      },
      "SymbolCatalog": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "title": "Symbols",
            "type": "array"
          }
        },
        "required": [
          "symbols"
        ],
        "title": "SymbolCatalog",
        "type": "object"
      },
      "UnsupportedStrategy": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "strategy": {
            "title": "Strategy",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "unsupported_strategy",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "detail",
          "strategy"
        ],
        "title": "UnsupportedStrategy",
        "type": "object"
      },
      "UnsupportedSymbol": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "unsupported_symbol",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "detail",
          "symbol"
        ],
        "title": "UnsupportedSymbol",
        "type": "object"
      },
      "UnsupportedTimeframe": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "unsupported_timeframe",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "detail",
          "timeframe"
        ],
        "title": "UnsupportedTimeframe",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Trading Terminal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/backtests": {
      "post": {
        "operationId": "backtests_backtests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/UnsupportedSymbol"
                    },
                    {
                      "$ref": "#/components/schemas/UnsupportedTimeframe"
                    },
                    {
                      "$ref": "#/components/schemas/UnsupportedStrategy"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidStrategyParameters"
                    }
                  ],
                  "title": "Response 400 Backtests Backtests Post"
                }
              }
            },
            "description": "The requested analysis inputs are not supported."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailable"
                }
              }
            },
            "description": "The market source cannot be read."
          }
        },
        "summary": "Backtests",
        "tags": [
          "analysis"
        ]
      }
    },
    "/backtests/strategies": {
      "get": {
        "operationId": "backtest_strategies_backtests_strategies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrategyDefinition"
                  },
                  "title": "Response Backtest Strategies Backtests Strategies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Backtest Strategies",
        "tags": [
          "analysis"
        ]
      }
    },
    "/candles": {
      "get": {
        "operationId": "candles_candles_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "in": "query",
            "name": "timeframe",
            "required": false,
            "schema": {
              "default": "1m",
              "title": "Timeframe",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleWindow"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/UnsupportedSymbol"
                    },
                    {
                      "$ref": "#/components/schemas/UnsupportedTimeframe"
                    }
                  ],
                  "title": "Response 400 Candles Candles Get"
                }
              }
            },
            "description": "The requested symbol or timeframe is not supported by this candle slice."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailable"
                }
              }
            },
            "description": "The market source cannot be read."
          }
        },
        "summary": "Candles",
        "tags": [
          "candles"
        ]
      }
    },
    "/client-events": {
      "post": {
        "operationId": "client_event_client_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Client Event",
        "tags": [
          "observability"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "system"
        ]
      }
    },
    "/ready": {
      "get": {
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Ready Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready",
        "tags": [
          "system"
        ]
      }
    },
    "/symbols": {
      "get": {
        "operationId": "symbols_symbols_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SymbolCatalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceUnavailable"
                }
              }
            },
            "description": "The market symbol catalog cannot be read from the source."
          }
        },
        "summary": "Symbols",
        "tags": [
          "candles"
        ]
      }
    },
    "/timeframes": {
      "get": {
        "operationId": "timeframes_timeframes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Timeframes Timeframes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Timeframes",
        "tags": [
          "candles"
        ]
      }
    }
  }
}